Return nil when a forecast is deleted
`ForecastHistory#set_value` needs to return either a `Forecast` or `nil`

Previously, in the particular case of setting the value of an original
forecast to zero (which results in deleting the original entry rather
than creating a zero value original entry) `set_value` was returning the
now-deleted entry.

This is not useful for our use cases, and was creating confusion when
using the return value of this method.

Return `nil` instead.

We’ve also added a test scenario for this specific situation to the
import spec.