Add Original Commitment Figure form step
We don't allow this to be edited after it's been set.

This has been a bit tricky, as I originally didn't want to have a
Commitment hanging around with no values on it, but creating a nested
entity with `fields_for` requires you to create a blank entity to be
used in the form here.

This is currently a mandatory step in the new activity journey, but
we'll make it optional for level B and C activities in a future commit,
so it won't be persisted at those levels with no values set on it.

Co-authored-by: Ynda Jas <ynda@dxw.com>