Add Original Commitment Figure form step
We don't allow this to be edited after it's been set.

This has been a bit tricky, as I originally didn't want to have a
Commitment hanging around with no values on it, but creating a nested
entity with `fields_for` requires you to create a blank entity to be
used in the form here. To get things to persist across requests when
they fail validation so you can render the error message, we're having
to save a blank Commitment on all Activities at this step.

Getting the validation errors to render here has been very difficult,
and I'm really not happy with the approach, but I _cannot_ see a way of
getting the created `Commitment` to persist between the initial page
load and re-loading the page with error messages without recreating the
`Commitment` in the error section of the `update` method. We should
definitely tidy this up in future but I've spent _days_ trying to figure
this out!

This is currently a mandatory step in the new activity journey, but
we'll make it optional for level B and C activities in a future commit.