Add `transaction_date` field to `Commitment`
In future, this will be populated with the `actual_start_date`, or the
`planned_start_date` depending on which is present, and will be output
in the IATI XML for an activity.

This will replace the `financial_year` and `financial_quarter` values we
currently display for historical commitments, as an Original commitment
figure should no longer be tied to a specific quarter, but represents
the expected spend for the entire lifetime of an Activity.