Add migration to set Commitment `transaction_date`
for historical Commitments.

In the past, Commitments appear to have been tied to a specific
financial year and quarter. This is no longer the case, as Commitments
should capture a committed spend over the entire lifetime of an Activity
(not just a single reporting period, but potentially multiple years).

Here we set the value to the `first_day_of_financial_period` for
existing Commitments that were created from a script in the past.