Refactor commitment importer & update validations
We no longer want to attribute a Financial Year or Financial Quarter to
Commitments, but will use a `transaction_date` instead.

The model now doesn't validate these fields, which will be removed in
the near future, but does validate the presence of a `transaction_date`.

Here we update the Commitment importer (currently only run as part of a
Rake task) to infer the `transaction_date` from the Activity's
`planned_start_date` or `actual_start_date` depending on which exists.
If neither exist (which, despite validations, seems to be the case for a
few Activities in the codebase), we fall back on the Activity's
`created_at` date.