Only generate IATI identifiers for ODA activities
Currently, we generate an IATI identifier for all activities.

However, we have had some feedback that this is confusing, and the IATI
identifier should be absent for non-ODA activities in the CSV download.

Only generate IATI identifiers for ODA activities. Add some coverage for the
`is_oda` and `transparency_identifier` attributes of the activity defaults.

We don't have any ISPF activities in production, so there aren't any existing
identifiers that need to be hidden.