Omit legacy fields for ISPF activities
Currently, we show the `intended_beneficiaries`, `recipient_country`, and
`recipient_region` legacy fields on the "details" page for all activities.

ISPF activities will not have the data associated with these legacy fields, so
we can remove the fields when the activity is of that type.

The XML download is unchanged. When an XML download is generated, the tags for
the legacy data are only created if the data is present. As ISPF activities
will never have any data for these field, there is no need for additional checks.