Hide SID truncation note on non-ODA activities
The purpose step of the activity form includes a note on both the title
and description fields about truncation of data when published to
Statistics in International Development. Non-ODA activities are not
published to this, so the note is irrelevant and confusing in that
context. This therefore hides the note when accessing the step on a
non-ODA activity