Fix error after `spring-watcher-listen` upgrade
Upgrading `spring-watcher-listen` from 2.0.1 to 2.1.0 introduces the
following error when `script/server` tries to run test database
migrations (`/Users/:username` changed to `~` in the code block)

```
~/.rbenv/versions/3.0.5/lib/ruby/gems/3.0.0/gems/spring-4.1.1/lib/spring/application.rb:100:in `block in preload': Spring reloads, and therefore needs the application to have reloading enabled.
Please, set config.cache_classes to false in config/environments/test.rb.
 (RuntimeError)
```

Doing what it says - setting `config.cache_classes` to `false` in
`config/environments/test.rb` fixes the server error, however it
introduces a deprecation warning when running tests, as below.
Tests pass and the full suite takes a similar time to run as on
`develop`

```
DEPRECATION WARNING: Initialization autoloaded the constant ApplicationHelper.

Being able to do this is deprecated. Autoloading during initialization is going
to be an error condition in future versions of Rails.

Reloading does not reboot the application, and therefore code executed during
initialization does not run again. So, if you reload ApplicationHelper, for example,
the expected changes won't be reflected in that stale Module object.

`config.autoloader` is set to `classic`. This autoloaded constant would have been unloaded if `config.autoloader` had been set to `:zeitwerk`.

In order to autoload safely at boot time, please wrap your code in a reloader
callback this way:

    Rails.application.reloader.to_prepare do
      # Autoload classes and modules needed at boot time here.
    end

That block runs when the application boots, and every time there is a reload.
For historical reasons, it may run twice, so it has to be idempotent.

Check the "Autoloading and Reloading Constants" guide to learn more about how
Rails autoloads and reloads.
```

A different fix for this was shared here:
https://github.com/rails/spring/issues/598#issuecomment-1268885973

Previously, we found that applying this fix may cause tests to run
slightly slower, however this no longer appears to be the case. All
tests pass in roughly the same timeframe as on `develop`, and there
is no deprecation warning. The server also loads fine with no issues
when accessing each of the pages linked in the navbar