Make sure the transparency indicator is nil for all non-ODA activities
Currently, the transparency indicator is set as `nil` in the ActivityDefaults
class.

However, these defaults are only applied to activities at the start of the UI
wizard, or through the import function.

This means that non-ODA activities that are created via the UI that do not have
a non-ODA parent, will have a transparency indicator. 

To fix, apply the defaults once non-ODA is selected as part of the wizard
journey as well.