Clear sessions during training DB sync
The previous commit addressed the case in which users encounter
and internal server error after we run the training database
sync script. Another scenario is that the user is able to access
the login screen and also able to reset their password, but
unable to log in before or after the reset. They see "Incorrect
Email or password." in the UI. This doesn't appear to produce a
server error

Like the internal server error, workarounds/fixes via manual
user intervention include using a different browser or an
incognito window, or deleting the `_roda_session` cookie. Unlike
the internal server error, it seems that clearing sessions on
the server fixes this one, so this adds a method/step to the
script to do this, using an existing Rake task