Put details page and summary linked activities behind a feature flag
We want to disable the linked activities feature for now, to be re-enabled in
future. We have had some feedback that linking is confusing to POs while
there's no ODA funds being spent.

This hides the linked activities row on the describe page and summary unless
the feature flag is active.