Nullify activity columns when activity is non-ODA
Certain columns should be null when an activity is non-ODA.

An activity can be created as non-ODA if it's parent activity is non-ODA, or if
it created through the import function, but can also change to non-ODA as part
of the UI wizard.

Add a before_save callback which sets any ODA-only attributes to nil when
is_oda is false, and the is_oda value has changed.

This also moves the nullification of transparency identifier to the model along
with the other null attributes.