Only errors on OCF upload if dates are present
Before this change, if you left out the `actual_` or `planned_`
`start_date` values, you'd get two errors: one for the missing date, one
for the invalid commitment (as these date values are used to set the
`transaction_date`).

Now that we only validate the associated commitment if one of these
values is present, we don't show the user a misleading, confusing error.