Decouple Import class from `activity_csv_columns.csv`
We want to be able to feature flag one of the CSV columns (activity linking).
However, this is difficult currently. Including the feature flag in the
`activity_csv_columns` initializer means that the effects of the feature flag
won't apply until the app has been restarted. Including the feature flag in the
current large Import class will be messy.

This decouples the rest of the `Import` class from the
`activity_csv_columns.csv` file, and provides an interface for common
operations.

Note: the `invalid_for_level_b_ispf_non_oda` and
`invalid_for_level_c_d_ispf_non_oda` are left as they were for now, but are
candidates for further refactoring.