List items in tab list links are presentational
The `ul` element has the role `tablist`. The specification says that it
must “own” `tab` elements.[1] We’ve chosen to give the links inside `li`
the role `tab`, rather than give `li` elements `tab` role.

Even though an “owned element” is defined as “any DOM descendant of the
element”, Axe tools raises a critical issue: “Element has children which
are not allowed” for the `ul` with role `tablist`, and “Required ARIA
parent role not present: tablist“ for the `a` elements with role `tab`,
suggesting that it was expecting direct child/parent elements rather
than going through the intermediate `li` element.

Elsewhere we fixed this by giving the `li` elements presentational roles

This also fixes 4 serious issues for the `li` elements: "List item does
not have a <ul>, <ol> parent element without a role, or a role="list”".

[1] https://www.w3.org/TR/wai-aria-1.1/#tablist