Remove mobile number reset from new user form
The new/edit user forms (which share a partial) have a field for
resetting the user’s mobile number. It says that “The user will have to
provide their mobile number on their next log in attempt”. In the new
user form, it’s unclear what not checking this would do for a new user -
would they be unable to provide their mobile number (on the first login,
when they need to provide one)?

Checking it leads to an internal server error when submitting the form,
so the correct way to submit the form is with this field unchecked

This removes the redundant and confusing field from the new user form