Signpost adding comments on actual upload
Some users try to upload comments via the actuals and refunds form
without an accompanying addition/change in values. This is not
supported, so here we're adding some signposting for how to do this on
the update view after an unsuccessful upload

The warning text that was added to the `new` view is now abstracted
into a partial with the warning text passed in. This is reused in the
`update` view