Make activity upload form IDs unique
TODO: unit test for `activity_csv_upload_file_field`

Oof. This was painful

This overwrites the default IDs for the `input` and `div.govuk-hint`
elements, the `for` attribute of the `label` element, and the
`aria-describedby` attribute of the `input` element, in order to address
critical accessibility issues highlighted by axe DevTools

I tried to use more of the design system native methods, but I couldn't
find a good way to get everything we need. In the end, I used
`GOVUKDesignSystemFormBuilder::Base.new` and then hacked the rest

I thought I was almost there, and then I ran axe DevTools and it listed
more critical errors, this time with aria. If you specify an
`aria-describedby` value, the design system appends it to the default
instead of overwriting it, and the default references an element that
doesn't exist (after we've overwritten the default hint IDs). This leads
to a critical issue. So in the end I chose to manually `.gsub` the
string that the field helper spits out to overwrite this attribute

`CGI.escapeHTML` is added in the controllers to address a Brakeman XSS
warning

The `TestHelper` class pattern in the spec is borrowed from
`spec/lib/roda_form_builder/form_builder_spec.rb` - this helps provide
access to a form builder