Make organisation IDs unique at /reports
For BEIS users, /reports has two tabs. On each tab, there is a heading
column for organisation names. The `th` elements use the organisation's
ID as their ID, but this is duplicated for each tab, which is
an accessibility issue. This adds the tab heading/report status
grouping ("current"/"approved") to the ID in order to create unique IDs