Only show IATI export with publishable activities
The link to download an IATI export for an organisationfund/level
combination of activities is shown if any activities exist for that
combination. However, when preparing the XML, activities are only
included if publishable to IATI (based on an activity's
`publish_to_iati` attribute). This means links can be generated for
exports that will contain no activities, which is invalid. This makes it
so that a link is only shown if publishable activities exist for the
given combination

I originally didn't notice the unit tests of `Iati::XmlDownload` and
added a feature test. When the unit tests failed, I updated them, but I
think the feature test might still be useful, since the other exports
have them, and it's fairly light touch

In updating the unit tests, I simplified some of the mocking to avoid
the chaos of combining mocking of chained calls with stubbed Active
Record relations, as well as cutting out mocking of the internal logic
of private methods. Instead, we now just mock the return value