Split spec into different user contexts
We originally had a "when the user is a BEIS user" context, then some
other free-dangling specs that defaulted to the user being a Partner
Organisation user which made it a bit tricky to know where to add new
tests here. We're about to add some tests for different functionality
depending on the user type, so it felt right to split them into two
contexts here.