Allow setting new commitment on existing activity
via the bulk upload.

Before this, any inclusion of the Commitment in the upload template
would cause the bulk upload to fail. We believe all users should be able
to at least initially set a Commitment on an existing activity if it
doesn't exist.

If there's already a commitment, we render an error message to let users
know they'll need to edit this via the user interface.