Use proper Haml instead of complex translation
We were using a complex HTML translation unnecessarily. To make the
translation more flexible for future usage, it's best to include HTML
only by exception. This moves most of the markup to the view, leaving
just the text content (plus mid-string link markup) as the
responsibility of the translation

The translation keys and structure have been updated to better reflect
our conventions

This also turns what was two paragraphs into two list items, in
preparation for adding a third in a later commit