Don't overwrite commitments when editing on error
Before this, if users edited a commitment and submitted an invalid
commitment value e.g. a string, it would overwrite the commitment and
set it to `nil` on the Activity when the validation failed. I'm
struggling to find any documentation for why this is the case, but I'm
assuming the Activity model saves itself when we call `save!`
unsuccessfully on the child association.

By putting this in a transaction, we roll back if `save!` fails.