Add script to create an audit for existing Budgets
We are using the Audited Gem to create revisions for Budgets.

The "Original" Budget will be the audit that is created when the Budget is
first created. However, while this works for new Budgets, we must create an
audit for all existing Budgets as well.

Audited does not provide an easy way to do this. This uses a private method of
the library, then updates the created_at attribute to match the updated_at
attribute of the Budget itself (updated_at represents the time that the Budget
was put in its current state and is our best guess at an original state).