Use correct params for put request
This was passing on `develop` due to us still calling the `update`
method in the `ActivityFormsController`, so the `:set_commitment?`
method was being called. Although it was technically ok for the params
to be incorrect for what we were testing here, this was surfaced when
updating the way we handle errors for Commitments, which is coming in a
future commit.