Handle errors at Commitment level
and refactor the way we build Commitments in the `Activity::Updater`'s
`set_commitment` method to follow example set in `BudgetCreator` and
other classes to make this easier to manage.

Before this, we'd add errors that occured when saving to the Activity,
rather than the Commitment so we could use the existing error flow in
the `ActivityFormsController#update` method. This felt a bit hacky, and
errors were swallowed and ignored.

We're about to add the ability to input financial values with
commas using the `ConvertFinancialValue` class, which throws an error
_before_ the `save!` method is called, meaning the unsaved
Commitment isn't available to be rendered with its errors.

We don't actually need to call `save!` with the `!` here and manually
add errors to the commitment. They'll be populated for us with the usual
`save` method so there's no need to `rescue` either.