Determine the IATI type of a Budget
We need to know whether a Budget is "original" or "revised".

This can be awkward, as a Budget object can either be a "revision" which is
returned by the Audited library, or the "live" Budget. Note: the language is
confusing here, but a "revision" returned from the Audited library is not
necessarily a "revision" in the application.

This adds methods to determine the IATI type of a Budget. It also adds some
additional helper methods, removes an unnecessary presenter method, and tidies
up some existing specs with a factory trait.