Install Audited for budget revisions
We want to add "revisions" to budgets.

This installs the Audited Gem which will enable us to record any changes to
budgets, along with who made them, and user provided comments.

Because we are using a UUID primary key for our Budget model, I have had to
update the column type of `auditable_id` in the generated migration.