Reorder deployment steps
In practice, we confirm the release before we put together the release
branch, to account for things such as any training sessions, reporting
windows during which the site might be busier etc.

We also need to think about dependencies beforehand, so they are ready
by the point we get to merging code.