Fix PO identifier validation and add tests
Previously `Activity::Import` was meant to check that when updating,
the partner organisation identifier was not present - presumably it's
not meant to be updatable. However, it was checking the heading
"Partner Organisation Identifier", when the template's heading (and test
object's key) was "Partner organisation identifier" (different case).
The case mismatch meant that this validation wasn't working properly -
it would always see `nil` because it was looking for a heading that
didn't exist. This fixes the validation and updates the test specs not
to try to update the field

This was partly covered by a test in
`spec/features/beis_users_can_upload_level_b_activities_spec.rb`,
but this was testing it using the wrong casing (i.e. not matching
the template), so we were getting a false negative (pass). This is
fixed, and an additional test has been added to the import spec in
line with similar tests. Future work needs to be done to refactor
the import specs more thoroughly, as discussed in fcd09c2