Fix flaky time travel spec
This spec is flaky in CI.

I'm not sure why, but sometimes when time travelling, the user gets signed
out, and the next assertion fails.

This re-authenticates the user after time travelling. It also freezes the time
in the correct before block, as previously it was accidentally part of a different
context.