Allow users to enter commas in OCF
Before this change, entering a comma in the OCF input field would throw
an error complaining that the user had not entered a real number.

We return early from the `set_commitment` method if the conversion to a
financial value fails here, as when editing an existing commitment, if we
rescue the `ConvertFinancialValue::Error`, it saves the original
commitment value and clears the `errors` array on the Commitment as it's
technically "valid" with its previous values.

We set the value of an invalid commitment to the erroneous value that
raised a `ConvertFinancialValue::Error` here to highlight the value that
threw the error - without this, we'd always be fetching the valid
commitment value and displaying it if one existed, which presented a
confusing user experience.