Accept commas in bulk upload for commitment value
This is a case where I feel duplication is preferable over trying to
force the behaviour in the `Activity::Updater` to work for both the
Activity Form journey _and_ the importer for the sake of keeping things
DRY.

This also nests commitment translations under the "commitment" now we
have multiple error messages

This addition surfaced an issue with the Importer class where any errors
at the Converter level will prevent model validations being displayed,
meaning the FSTC being blank wasn’t picked up as an error and the test
began failing.

We’ll want to improve this behaviour if possible, but to prevent scope
creep I've removed that assertion to allow all tests to pass without
having to fix that issue.

Ideally, we should display both model and converter-level errors in the
bulk upload. This will be handled in [this future
task](https://trello.com/c/vr1rInEO/2891-improve-activity-importer-validation)