Add inclusion validators to Activity attributes
The results of a recent pen test raised that various attributes of an Activity
can be set to arbitrary values.

This adds inclusion validation for the attributes.

Validation is kept loose where we use codelists, we only check whether a value
is in the corresponding codelist, and not whether the value is current. This
means that whilst a user could potentially enter an incorrect code, no
malicious values can be entered. Because of this, no older activities will be
invalidated (e.g., Activities that have a withdrawn `sector_category` code).

Some validators use helper methods so that the options are consistent between
the validation and radio options; whilst not ideal, this ensures the code is
DRY.