Prevent overflow of tables
The results of a recent accessibility audit raised that the content of some
tables overflows the viewport width without the ability to scroll right or
adapt the content to fit inside the available space.

This takes the approach found here
https://github.com/DFE-Digital/register-trainee-teachers/pull/2217, which in
turn references the table on the GOV.UK Cookies page
https://www.gov.uk/help/cookie-details.

Instead of applying the styles to the table itself, we apply them to a
container as suggested by the accessibility audit.

a