Ask user to confirm activity deletion
After clicking on "Delete this activity" from an activity page, the user
should be asked to confirm this action

Initially I'd introduced an `ActivityDeletionsController` in this commit
after brief consultation with a couple of colleagues on how to represent
a "confirm destroy" step in a controller action. However, on further
dicussion with the RODA devs and some research that uncovered a common
`confirm_destroy` action, this has been refactored to bring the action
into the activities controller. A later commit will add the `destroy`
action to the same controller

While many of our controller tests set up complex context structures,
here the policy dependency of the controller is mocked so that we are
only testing the controller itself. The different combinations of
activity level, user organisation and titled-ness are the concern of the
policy, while the controller cares only about the resulting
authorised/unauthorised binary. We don't stub
`prepare_default_activity_trail` and its result isn't tested here (it is
in the feature spec), but we do stub a user with a `service_owner?`
method to allow it to run without errors