Delete activity when request is confirmed
This is the final step in the activity deletion journey. When a BEIS
user confirms their request to delete an activity, the activity, its
associations and its children should all be deleted

This is the first use of the GOV.UK notification banner component, which
I've extracted into a partial specific to this context. The partial is
called in two different views which will be reached dependent on whether
the deleted activity was a programme or project/third-party project. The
feature tests therefore cover deleting both a programme activity, which
will lead to `index_beis`, and a project activity, which will lead to
`index`

In the controller spec, I would have liked to have included the
assertion `expect(activity).to receive(:destroy)` or similar with
`have_received` to keep things constrained to what the controller
itself is doing, but this didn't work. Instead, we check that there is
no longer an activity with that ID, similar to the feature test