Use safe_join in accessible links
Using `html_safe` on text that can contain user input is not advised, as
any user input has to be considered potentially malicious.

Rails already escapes any strings not explicitly marked as safe. When
using `html_safe`, it becomes our responsibility to ensure the text is
indeed safe to display as HTML.

Use `safe_join` instead, so the `span` is still displayed properly, but
any unsafe text is still escaped.

Additionally, the specs for the `a11y_action_link` method were more or
less the same as the implementation, so I've made them more explicit
about the contents of what they're testing.