Use `Time.current` instead of `Time.now`
in test setup.

This test began failing after the switch to British Summer Time.

`Time.now` uses the system time (which is BST) while `Time.current` uses
the time in the current zone (defaulting to UTC in Rails).

An alternative to this would have been to set the `config.time_zone` in
`application.rb`. Although we don't have users outside the UK who might
be confused by this, it's still best practice to avoid setting an
application-wide time zone.

For more information, [this blog
post](https://thoughtbot.com/blog/its-about-time-zones) was helpful at
explaining how to work best with timezones in Rails.