Add task to generate a CSV of Activities for annual fund impact metrics
We want to be able to generate a CSV of activities that will be used to
pre-populate an annual collaborative spreadsheet of fund impact metrics.

This task will be used to capture a snapshot of activities once a year in
July.

The CSV will be created in the `tmp` folder. We have documentation describing
how to download files when connected remotely to an instance of the application.