PO users can view implementing orgs
PO users do not get access to the View page for each organisation yet,
but this is planned for future iterations of this page.

Rather than having two separate index pages, one for service owners and
one for PO users, I've chosen to use the same partial for the org table,
and show or hide the Actions column accordingly.