Remove linked activity ID before destruction
A recent change introduced functionality to delete untitled activities.
When such activities had a linked activity, the remaining activity would
show the option to add a new linked activity, but it would still hold
the deleted activity's ID as its linked activity ID in the database
until it was either explicitly marked as having no link or linked to a
different activity

While not thoroughly tested, this could potentially cause errors if any
functionality was built around linked activity IDs (and an assumption
that if one existed, it belonged to an existing activity)

This ensures the linked activity ID of the remaining activity in a
linked pair is removed before an activity is deleted. Because we're
using a `before_destroy` hook, this applies to all children of
activities the user requests to be deleted (they're destroyed in the
process)