Fix XML policy for funds and programmes
Fund and programme activities are not meant to be shown as XML on their
own. (Rendering fund-wide XMLs is handled at organisation level, through
the exports functionality.)

A recent change, intended to consolidate XML policies, has inadvertently
resulted in "Download as XML" buttons appearing on funds and programmes
for BEIS users.

While not revealing any information BEIS users don't already have access
to, this change is confusing and not necessary.

Additionally, there were no specs that would have caught this specific
regression.

I've chosen to rename the `download?` method to `show_xml?` to be
consistent with the exports organisation policy, and to be specific
about the target of the policy.