Make sure users can sign out if JavaScript is disabled
Rails uses JavaScript to intercept requests from links that are given a
`method`, which allows us to use links for methods including DELETE. When
JavaScript is disabled, this falls back to a GET request.

Currently, this means that users see an error page when attempting to sign out
with JavaScript disabled, as Devise is expecting a DELETE request.

Devise allows us to specify the method used to sign out, so we can update this
to GET.