Allow valid actual/refund without deleting invalid rows
The actuals/refunds upload template is prepopulated with rows for all
the activities to which you can add values. Previously, if you didn't
want to provide values for all activities (e.g. if you just wanted to
add values for one activity), you would have to remove all the remaining
rows, since the default values are invalid

This makes it so that if there's at least one valid row and the only
errors are default value errors, the valid row is imported and the
invalid rows are ignored/discarded. This saves users the effort of
manually deleting rows they don't care about

Rows with other errors will still prevent the import from completing,
and the actual: zero, refund: blank error will still show if there are
other errors and/or no valid rows

The `has_only_default_value_errors?` method relies on the error column
being the string "Actual Value/Refund Value" and the message being a
specific translation. This doesn't feel as resilient as I'd like, but
hopefully the tests are sufficient enough to catch a breaking change

I would have preferred to have access to the type hash created in
`Converter#validate_values`, however getting this all the way up to the
base `Import` class would likely mean a more significant refactor,
modifying both the `Error` `Struct` and the iteration over
`importer.errors` in `import_row`. If we added a `value_type` attribute
to `Error`, this wouldn't be relevant to errors from non-value columns,
which would make things a little messy