Reject actual/refund import when invalid with comment
The previous commit made it so that rows with (invalid) default values
would be ignored if there was one valid row and no other errors.
However, if a user added a comment to one of the rows with (invalid)
default values, this would be discarded without alerting the user

This makes it so that the erroring happens as usual if
`invalid_with_comment` is `true`

I think the error messaging could be clearer, given the error table
would then say "Actual can't be zero when refund is blank" but
technically that's now okay as long as there's no comment and another
row is valid. However, in this context we do have the signposting via
the warning component ("Comments can only be added via the bulk upload
if..."). It's probably fine for now, and one to test with users in the
next phase