WIP add fixed data migration
After syncing production data with my local environment, the original
data migration file didn't work for me. I've fixed most of the issues
and simplified it by using `CreateBudget`, however the `Report` model
has the following validation:

```ruby
validates_presence_of :report, unless: -> { parent_activity&.organisation&.service_owner? }
```

BEIS/DSIT is the only 'service owner' organisation and Level D
activities don't belong to the BEIS/DSIT `organisation`, so a report
is needed

There will only be an `editable_report_for_activity` (used in
`CreateBudget`) if there's a current open report, and we probably don't
want to add these to the currently open report in any case, since that
will be for the current financial year and quarter

We could connect these budgets to an old report if one exists, but
since reports belong to a specific financial quarter, we'll either need
DSIT to provide a financial quarter for all of these budgets, or we
could assign them to a report in the first financial quarter for the
given financial year (I think they've suggested this for some other
purpose before)

We *might* need to deal with there being no report for the financial
quarter in either case

But it would be good to clarify if they want these adding to historical
reports as these are not usally modified once approved

Alternatively, I think there could be a way to persist these without a
report by bypassing validations, but we'd have to be sure DSIT are
happy having budgets on level D activities that aren't attached to a
report, since the schema isn't designed for this scenario - it only
allows report-less budgets at level B

If we are attaching these to a report, we could probably still use
`CreateBudget` to do most of the work, then take its `.object` and add
a (non-editable) report to it, then try saving again