Change error handling in CmdRunner to handle stderr errors
The training_db_sync script  does not break when errors are thrown while running commands with Ruby module Open3.

This edit changes the error handling in the script to check for errors present in stderr. This allows us to capture any such errors, and raise an exception if any are present. If errors are present in stderr while capturing data from source; copying data; or loading data to the destination, the script will delete sessions and tempfiles before exiting.