Move test expectation to its correct context
In the new infrastructure, there is only a private bucket, so we need to
isolate all the code pathways that use a public one, to ensure that the
application works fully with a private bucket.

Changing the generic context for all specs to a `use_public_bucket`
value of `false` resulted in one spec failing. A closer look revealed
that it's failing because it's not meant to work for the private bucket,
so moving it to its correct context is the right thing to do.