Amend report validation to include is_non_oda
Amend the report validation (enforced as Rails model validation and as
database unique index) to take into account the ODA type of the report.

This allows non-ISPF reports (whose `is_non_oda` value is `nil`) to
function as usual, and it allows ISPF to have one ODA and one non-ODA
report editable at the same time for each partner organisation.