Add spec for invoking policy with a symbol
Generally we prefer applying the policies to a specific instance (such
as a project or third-party project), because then the policy has the
most information (such as the organisation responsible for the activity).

However, in the case of the IATI redaction, BEIS/DSIT users can perform
this action on any given project and third-party project, so it is more
efficient to check against a symbol rather than a specific project. This
policy is also used to decide whether a table header will be shown or
not, and there is no specific project at the time of rendering the table
header, so the symbol check is necessary.

This usage was not tested in the policy spec before.