Remove redundant specs
Now that [ThirdParty]ProjectPolicy inherit most of their behaviour from
ActivityPolicy, we can remove the redundant specs.

In a few cases, the activity policy spec wasn't testing the full range
of actions, namely the `new` action, so we've added the test wherever it
was missing.

In a few other cases, the project policy specs were erroneous, as in the
case of edit and update. The activity policy specs cover all the
configurations such as whether a PO user is looking at other org's
activities, and whether there is an editable report or not.